Recent observations at hotspots

Returns the most recent sighting information reported in a given vector of hotspots.

ebirdhotspot(
  locID,
  species = NULL,
  back = NULL,
  max = NULL,
  locale = NULL,
  provisional = FALSE,
  sleep = 0,
  key = NULL,
  ...
)

Arguments

locID
(required) Vector containing code(s) for up to 10 regions of interest; here, regions are the locIDs of hotspots. Values that are not valid or are not hotspots are ignored.

species
Scientific name of the species of interest (not case sensitive). Defaults to NULL, in which case sightings for all species are returned. See eBird taxonomy for more information: http://ebird.org/content/ebird/about/ebird-taxonomy

back
Number of days back to look for observations (between 1 and 30, defaults to 14).

max
Maximum number of result rows to return in this request (between 1 and 10000, defaults to all)

locale
Language/locale of response (when translations are available). See http://java.sun.com/javase/6/docs/api/java/util/Locale.html (defaults to en_US)

provisional
Should flagged records that have not been reviewed be included? (defaults to FALSE)

sleep
Time (in seconds) before function sends API call (defaults to zero. Set to higher number if you are using this function in a loop with many API calls).

key
eBird API key. You can obtain one from https://ebird.org/api/keygen. We strongly recommend storing it in your .Renviron file as an environment variable called EBIRD_KEY.

...
Curl options passed on to GET

Value

A data.frame containing the collected information:

"comName": species common name

"howMany": number of individuals observed, NA if only presence was noted

"lat": latitude of the location

"lng": longitude of the location

"locID": unique identifier for the location

"locName": location name

"locationPrivate": TRUE if location is not a birding hotspot

"obsDt": observation date formatted according to ISO 8601 (e.g. 'YYYY-MM-DD', or 'YYYY-MM-DD hh:mm'). Hours and minutes are excluded if the observer did not report an observation time.

"obsReviewed": TRUE if observation has been reviewed, FALSE otherwise

"obsValid": TRUE if observation has been deemed valid by either the automatic filters or a regional viewer, FALSE otherwise

"sciName" species' scientific name

References

http://ebird.org/

Author

Rafael Maia rm72@zips.uakron.edu

Examples

if (FALSE) {
ebirdhotspot(locID=c('L99381','L99382'), species='larus delawarensis')
ebirdhotspot('L99381', max=10, provisional=TRUE)
}